// src/app/nw-data/src/types.ts
export interface HouseItems {
  AttributionId: string;
  AudioCreated: string;
  AudioDestroyed: string;
  AudioPickup: string;
  AudioPlace: string;
  AudioUse: string;
  BindOnPickup: boolean;
  ColorFamilies: string;
  ConfirmBeforeUse: boolean;
  ConsumeOnUse: boolean;
  CraftingRecipe: string;
  'DEV-FurnitureSet': string;
  DeathDropPercentage: number;
  Description: string;
  ExcludeFromGame: number;
  ForceRarity: number;
  HiResIconPath: string;
  HouseItemID: string;
  HousingStatusEffect: string;
  'HousingTag1 Placed': HousingTag[];
  'HousingTag2 Points': HousingTag[];
  'HousingTag3 Limiter': HousingTag[];
  'HousingTag5 Buffs': HousingTag[];
  HousingTags: HousingTag[];
  'HowToOptain (Primarily)': string;
  IconPath: string;
  InteractionAnimationID: string;
  IsEntitlement: boolean;
  IsRepairable: boolean;
  IsSalvageable: boolean;
  ItemRarity: string;
  ItemType: ItemType;
  ItemTypeDisplayName: string;
  MaxPotentialPoints: string | number;
  MaxStackSize: number;
  Name: string;
  Nonremovable: boolean;
  Notes: string;
  PlacementGridDisplaySize: string;
  PointModifier: number;
  PrefabPath: string;
  'Primary Color': string;
  RankingPoints: string | number;
  RankingPointsDuplicateLimit: number;
  RankingPointsNegativeLimit: number;
  RepairDustModifier: number;
  RepairRecipe: string;
  SalvageGameEventID: string;
  SalvageRecipe: string;
  SalvageResources: boolean;
  'Secondary Color': string;
  StorageBonus: number;
  'Tertiary Color': string;
  Tier: number;
  TradingCategory: TradingCategory;
  TradingFamily: TradingFamily;
  TradingGroup: TradingGroup;
  UIHousingCategory: string;
  UiItemClass: string;
  Weight: number;
}
// ...
export interface MasterItemDefinitions {
  AcquisitionNotificationId: number
  ArmorAppearanceF: string
  ArmorAppearanceM: string
  AttributionId: string
  AudioCreated: unknown
  AudioDestroyed: string
  AudioPickup: unknown
  AudioPlace: unknown
  AudioUse: unknown
  BindOnEquip: number
  BindOnPickup: number
  CanHavePerks: number
  CanReplaceGem: number
  CanRollPerkOnUpgrade: string
  ConfirmBeforeUse: number
  ConfirmDestroy: number
  ConsumeOnUse: number
  ContainerGS: number
  ContainerLevel: number
  CraftingRecipe: string
  DeathDropPercentage: number
  Description: string
  DestroyOnBreak: number
  Durability: number
  DurabilityDmgOnDeath: number
  EventId: string
  ExceedMaxIndex: string
  ExceedMinIndex: number
  ExclusivelyForWarCampTier: number
  ExtraBonusItemChance: string
  FFAConvertedItemId: string
  ForceRarity: number
  GearScoreOverride: number
  GrantsHWMBump: number
  HeartgemRuneTooltipTitle: string
  HeartgemTooltipBackgroundImage: string
  HiResIconPath: string
  HideFromRewardOpenPopup: number
  HideInLootTicker: number
  HousingTags: string
  IconPath: string
  IgnoreHWMScaling: number
  IgnoreNameChanges: number
  IgnoreParentColumns_DVT: string
  IngredientBonusPriApry: string
  IngredientBonusPriAugy: string
  IngredientBonusPriMayy: string
  IngredientBonusPrimary: number
  IngredientBonusSecondary: number
  IngredientCategories: string[]
  IngredientGearScoreBaseBonus: number
  IngredientGearScoreMaxBonus: string | number
  IsMissionItem: number
  IsRepairable: number
  IsRequiredItem: number
  IsSalvageable: boolean
  IsUniqueItem: number
  ItemClass: ItemClass[]
  ItemID: string
  ItemStatsRef: string
  ItemType: ItemType
  ItemTypeDisplayName: string
  MannequinTag: string
  MaxCraftGS: number
  MaxGearScore: number
  MaxStackSize: number | string
  MinGearScore: number
  MountItemAppearance: string
  Name: string
  NoBindOnPickupChance: number
  Nonremovable: number
  NotDroppable: boolean
  NotDroppableNotificationOverride: string
  Notes: string
  ObtainableReleaseVersion: string
  OnEquipObjectiveId: string
  ParentItemId_DVT: string
  Perk1: string
  Perk2: string
  Perk3: string
  Perk4: string
  Perk5: string
  PerkBucket1: string
  PerkBucket2: string
  PerkBucket3: string
  PerkBucket4: string
  PerkBucket5: string
  PerkSlot1: string
  PerkSlot2: string
  PerkSlot3: string
  PerkSlot4: string
  PerkSlot5: string
  PrefabPath: string
  RepairDustModifier: number
  RepairGameEventID: string
  RepairRecipe: string
  RepairTypes: string
  RequiredExpansionId: string
  RequiredLevel: number
  RequiredTradeskillRank: string
  RequiredWorldTags: string
  RetiredInReleaseVersion: string
  SalvageAchievement: string
  SalvageEntitlementId: string
  SalvageGameEventID: string
  SalvageGuaranteedPerkCount: number
  SalvageLootTags: string[]
  SalvageResources: number
  ScheduleId: string
  SoundTableID: string
  Tier: number
  TradingCategory: TradingCategory
  TradingFamily: TradingFamily
  TradingGroup: TradingGroup
  UiItemClass: string
  UseMagicAffix: number
  UseMaterialAffix: number
  UseTypeAffix: number
  WarboardDepositStat: string
  WarboardGatherStat: string
  WeaponAccessory: string
  WeaponAppearanceOverride: string
  Weight: number
}
// ...
export type HousingTag =
  | '#NAME?'
  | 'AllCombat'
  | 'AllCrafting'
  | 'AllGathering'
  | 'CombatAncientBonus'
  | 'CombatAngryEarthBonus'
  | 'CombatBeastBonus'
  | 'CombatCorruptedBonus'
  | 'CombatLostBonus'
  | 'CombatVarangianBonus'
  | 'CraftingAlchemyBonus'
  | 'CraftingBlacksmithingBonus'
  | 'CraftingCookingBonus'
  | 'CraftingEngineeringBonus'
  | 'CraftingOutfittingBonus'
  | 'GatheringFishingBonus'
  | 'GatheringHarvestingBonus'
  | 'GatheringLoggingBonus'
  | 'GatheringMiningBonus'
  | 'GatheringSkinningBonus'
  | 'IsAnimal'
  | 'IsBed'
  | 'IsBooksPaper'
  | 'IsCabinetShort'
  | 'IsCabinetTall'
  | 'IsChairDining'
  | 'IsChairOther'
  | 'IsCurtain'
  | 'IsDishes'
  | 'IsLargeDecor'
  | 'IsLighting'
  | 'IsMediumDecor'
  | 'IsPainting'
  | 'IsPlaceSetting'
  | 'IsRug'
  | 'IsShelfWall'
  | 'IsShelves'
  | 'IsSmallDecor'
  | 'IsStorage'
  | 'IsStoveOven'
  | 'IsTableDining'
  | 'IsTableOther'
  | 'IsTrophyBuff'
  | 'IsUnique'
  | 'IsVeg'
  | 'LimiterAnimatedRigged'
  | 'LimiterAnimatedTimelineInteract'
  | 'LimiterAnimatedTimelineLooping'
  | 'LimiterLighting'
  | 'LimiterStorage'
  | 'LimiterTrophyBuff'
  | 'MedleyFaireBonus'
  | 'NoStacking'
  | 'OnCeiling'
  | 'OnFloor'
  | 'OnFurniture'
  | 'OnWall'
  | 'OtherFoodTimeBonus'
  | 'OtherLootLuckBonus'
  | 'PvPXPBonus'
  | 'SeasonXPBonus'
// ...
export type ItemType =
  | 'Ammo'
  | 'Armor'
  | 'Blueprint'
  | 'Consumable'
  | 'Currency'
  | 'Dye'
  | 'HousingItem'
  | 'Lore'
  | 'MountDye'
  | 'Resource'
  | 'ThrowableItem'
  | 'Weapon'
// ...
export type TradingCategory =
  | 'Apparel'
  | 'Furnishings'
  | 'Resources'
  | 'Tools'
  | 'Utilities'
  | 'Weapons'

export type TradingFamily =
  | '1Handed'
  | '2Handed'
  | 'AlchemyResources'
  | 'Amulets'
  | 'ArcanaResources'
  | 'ArmorHeavy'
  | 'ArmorLight'
  | 'ArmorMedium'
  | 'Azoth'
  | 'Bags'
  | 'Bait'
  | 'Coating'
  | 'Components'
  | 'CookingIngredients'
  | 'CookingRecipes'
  | 'DyeIngredients'
  | 'Dyes'
  | 'Earrings'
  | 'Enhancements'
  | 'EventConsumables'
  | 'FishingTreasure'
  | 'FoodAttribute'
  | 'FoodRecovery'
  | 'FoodTradeskill'
  | 'GatheringTools'
  | 'HousingTrdF_Decorations'
  | 'HousingTrdF_Furniture'
  | 'HousingTrdF_Misc'
  | 'HousingTrdF_Trophies'
  | 'Instruments'
  | 'Magical'
  | 'MusicSheets'
  | 'Potion'
  | 'RawResources'
  | 'RefinedResources'
  | 'Rings'
  | 'RuneglassFamily'
  | 'SchematicsFurnishing'
  | 'Shields'
  | 'SpecialResources'
  | 'Tincture'

export type TradingGroup =
  | '1hSickle'
  | '1hSkinningKnife'
  | '2hAxe'
  | '2hPick'
  | 'AlchemyMagical'
  | 'AlchemyMedicinal'
  | 'AlchemyOffensive'
  | 'AlchemyProtective'
  | 'ArcanaAir'
  | 'ArcanaDeath'
  | 'ArcanaEarth'
  | 'ArcanaFire'
  | 'ArcanaLife'
  | 'ArcanaSpirit'
  | 'ArcanaWater'
  | 'AttributeItem'
  | 'AzothVial'
  | 'BagA'
  | 'Bait'
  | 'BaitFresh'
  | 'BaitSalt'
  | 'Beeswax'
  | 'Berry'
  | 'Block'
  | 'Blunderbuss'
  | 'Bows'
  | 'Charcoal'
  | 'Cloth'
  | 'CoatingAeternian'
  | 'CoatingAncient'
  | 'CoatingAngryEarth'
  | 'CoatingBeast'
  | 'CoatingCorrupted'
  | 'CoatingLost'
  | 'Conditional'
  | 'ConsumableLuck'
  | 'CookingComponents'
  | 'CraftedAmulet'
  | 'CraftedEarring'
  | 'CraftedRing'
  | 'CraftingComponents'
  | 'CraftingConsumables'
  | 'CraftingPatterns'
  | 'CutGem'
  | 'Defensive'
  | 'DroppedAmulet'
  | 'DroppedEarring'
  | 'DroppedRing'
  | 'Drums'
  | 'DyeBlue'
  | 'DyeBrown'
  | 'DyeCyan'
  | 'DyeGreen'
  | 'DyeGrey'
  | 'DyeMagenta'
  | 'DyeOrange'
  | 'DyePurple'
  | 'DyeRed'
  | 'DyeTurquoise'
  | 'DyeYellow'
  | 'Egg'
  | 'Feather'
  | 'Fiber'
  | 'FishLarge'
  | 'FishMedium'
  | 'FishSmall'
  | 'FishingPoles'
  | 'FishingSalvage'
  | 'Flail'
  | 'Flutes'
  | 'FoodArcanist'
  | 'FoodArmorer'
  | 'FoodAttribute'
  | 'FoodCON'
  | 'FoodDEX'
  | 'FoodEngineer'
  | 'FoodFOC'
  | 'FoodHarvester'
  | 'FoodHealthRecovery'
  | 'FoodHunter'
  | 'FoodINT'
  | 'FoodJewelcrafter'
  | 'FoodLumberjack'
  | 'FoodMana'
  | 'FoodMiner'
  | 'FoodSTR'
  | 'FoodWeaponsmith'
  | 'Fruit'
  | 'GauntletIce'
  | 'GauntletVoid'
  | 'Grain'
  | 'GreatAxes'
  | 'GreatSword'
  | 'Guitars'
  | 'Gunpowder'
  | 'Hatchets'
  | 'Honey'
  | 'HousingTrdG_Animals'
  | 'HousingTrdG_Beds'
  | 'HousingTrdG_BooksPaper'
  | 'HousingTrdG_BuffCombat'
  | 'HousingTrdG_BuffCrafting'
  | 'HousingTrdG_BuffGathering'
  | 'HousingTrdG_BuffOther'
  | 'HousingTrdG_Cabinets'
  | 'HousingTrdG_Chairs'
  | 'HousingTrdG_Curtains'
  | 'HousingTrdG_DecorOther'
  | 'HousingTrdG_Dishes'
  | 'HousingTrdG_Lighting'
  | 'HousingTrdG_MiscOther'
  | 'HousingTrdG_Paintings'
  | 'HousingTrdG_Rugs'
  | 'HousingTrdG_Shelves'
  | 'HousingTrdG_Storage'
  | 'HousingTrdG_Stoves'
  | 'HousingTrdG_Tables'
  | 'HousingTrdG_Veg'
  | 'Ingot'
  | 'IngotPrecious'
  | 'InstrumentComponents'
  | 'Leather'
  | 'Mandolins'
  | 'Meat'
  | 'Muskets'
  | 'Nut'
  | 'Offensive'
  | 'Oil'
  | 'Ore'
  | 'OrePrecious'
  | 'Pigments'
  | 'PotionElemDefense'
  | 'PotionEncumbrance'
  | 'PotionFocus'
  | 'PotionHealth'
  | 'PotionMana'
  | 'PotionPurifyB'
  | 'PotionWellness'
  | 'Rapiers'
  | 'RawGem'
  | 'Rawhide'
  | 'RecipeFurnishing'
  | 'Recipes'
  | 'RefiningComponents'
  | 'RuneglassAmber'
  | 'RuneglassAmethyst'
  | 'RuneglassAquamarine'
  | 'RuneglassCarnelian'
  | 'RuneglassDiamond'
  | 'RuneglassEmerald'
  | 'RuneglassEmpty'
  | 'RuneglassJasper'
  | 'RuneglassMalachite'
  | 'RuneglassMoonstone'
  | 'RuneglassOnyx'
  | 'RuneglassOpal'
  | 'RuneglassRuby'
  | 'RuneglassSapphire'
  | 'RuneglassTopaz'
  | 'Salvage'
  | 'Seasonings'
  | 'ShieldsKite'
  | 'ShieldsRound'
  | 'ShieldsTower'
  | 'SlotChest'
  | 'SlotFeet'
  | 'SlotHands'
  | 'SlotHead'
  | 'SlotLegs'
  | 'Spears'
  | 'StavesFire'
  | 'StavesLife'
  | 'Stone'
  | 'Swords'
  | 'Timber'
  | 'TrinketComponents'
  | 'UprightBass'
  | 'Utility'
  | 'Vegetable'
  | 'WarHammers'
  | 'Water'
  | 'Wood'